Replace internal route with guild route and update syncProAccounts endpoint to syncProAccounts for improved clarity and functionality.
